+++ Application/__init__.py
@@ -1,5 +1,4 @@
 import os
-from . import db
 from flask import Flask
 
 
@@ -8,7 +7,8 @@ def create_app(test_config=None):
     app = Flask(__name__, instance_relative_config=True)
     app.config.from_mapping(
         SECRET_KEY='dev',
-        DATABASE=os.path.join(app.instance_path, 'Project.sqlite'),
+        #DATABASE= 'database.db',
+        
     )
 
     if test_config is None:
@@ -23,9 +23,6 @@ def create_app(test_config=None):
         os.makedirs(app.instance_path)
     except OSError:
         pass
-
-    db.init_app(app)
-
 
 
     return app
