+++ Application/db.py
@@ -1,41 +1,24 @@
-import sqlite3
-
-import click
-from flask import current_app, g
-from flask.cli import with_appcontext
+from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
 
 
-def get_db():
-    if 'db' not in g:
-        g.db = sqlite3.connect(
-            current_app.config['DATABASE'],
-            detect_types=sqlite3.PARSE_DECLTYPES
-        )
-        g.db.row_factory = sqlite3.Row
+'''
+db = SQLAlchemy() 
+app = Flask(__name__)
+app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///users.db'
 
-    return g.db
-
-
-def close_db(e=None):
-    db = g.pop('db', None)
-
-    if db is not None:
-        db.close()
-
-def init_db():
-    db = get_db()
-
-    with current_app.open_resource('schema.sql') as f:
-        db.executescript(f.read().decode('utf8'))
-
-def init_app(app):
-    app.teardown_appcontext(close_db)
-    app.cli.add_command(init_db_command)
-
+def initialization(app):
+	app.config.setdefault('SQLALCHEMY_TRACK_MODIFICATIONS', False)
+	db.init_app(app)
     
-@click.command('init-db')
-@with_appcontext
-def init_db_command():
-    """Clear the existing data and create new tables."""
-    init_db()
-    click.echo('Initialized the database.')
+
+
+
+class User(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    username = db.Column(db.String(120), unique=True, nullable=False)
+    password = db.Column(db.String(120), unique=False, nullable=False)
+    role = db.Column(db.String(120), unique=False, nullable=False)
+
+    def __repr__(self):
+        return '<User %r>' % self.username'''
