--- Application/schema.sql
@@ -1,9 +1,0 @@
-DROP TABLE IF EXISTS user;
-
-
-CREATE TABLE user (
-  id INTEGER PRIMARY KEY AUTOINCREMENT,
-  username TEXT UNIQUE NOT NULL,
-  password TEXT NOT NULL,
-  role TEXT NOT NULL
-);
