+++ Application/templates/base.html
@@ -4,7 +4,7 @@
 <nav>
     {% if session['username'] %}
       <span>Hello {{ session['username'] }}  |  </span>
-      <a href="{{ url_for('logout') }}">Log Out</a>
+      <!--<a href="{{ url_for('logout') }}">Log Out</a>-->
     {% else %}
     <a href="{{ url_for('login') }}">Log in</a>
   {% endif %}
