+++ config.py
@@ -1,0 +1,1 @@
+SECRET_KEY='dev'
