+++ main.py
@@ -1,5 +1,6 @@
 import Application
+
 app = Application.create_app()
 
 if __name__ == '__main__':
-	app.run(host='127.0.0.1',port=8080,debug=True)
+	app.run()
