+++ test.py
@@ -1,0 +1,5 @@
+import json
+
+users = json.load(open('users.json'))
+for i in users["users"]:
+    print(i)
